+++ src/libsodium/crypto_pwhash/scryptsalsa208sha256/sse/pwhash_scryptsalsa208sha256_sse.c
@@ -324,7 +324,7 @@ escrypt_kdf_sse(escrypt_local_t * local,
         return -1;
     }
 #endif
-    if ((uint64_t)(r) * (uint64_t)(p) >= (1 << 30)) {
+    if ((uint64_t)(r) * (uint64_t)(p) >= ((uint64_t) 1 << 30)) {
         errno = EFBIG;
         return -1;
     }
